# html/docs/openapi.yaml
openapi: 3.1.1
info:
  title: API Médicaments FR
  description: |
    API haute performance pour accéder aux données des médicaments français issus de la Base de Données Publique des Médicaments (BDPM).

    **Caractéristiques principales :**
    - Données officielles mises à jour automatiquement 2 fois par jour (6h et 18h)
    - Accès gratuit avec rate limiting équitable (1000 tokens initiaux, recharge à 3 tokens/sec)
    - Données intègres et non altérées depuis la source BDPM
    - Endpoints v1 modernes + endpoints legacy (dépréciés, suppression prévue le 31 juillet 2026)

    **Remarque :** Les endpoints legacy sont toujours fonctionnels mais seront supprimés. Veuillez migrer vers les endpoints v1.

    **Ressources utiles:**
    - [Guide de migration vers v1](https://github.com/Giygas/medicaments-api/blob/main/docs/MIGRATION.md)
    - [Base de données BDPM](https://base-donnees-publique.medicaments.gouv.fr)

  version: 1.1.0
  contact:
    name: GitHub Repository
    url: https://github.com/giygas/medicaments-api
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: https://medicaments-api.giygas.dev/
security: []
...
paths:
  /v1/medicaments:
    get:
      summary: Obtenir des médicaments (v1)
      description: |
        Recherche et récupération de médicaments avec différents paramètres de requête.
        Seul un paramètre est autorisé à la fois.
      tags:
        - Médicaments (v1)
      parameters:
        - $ref: "#/components/parameters/QueryPage"
        - $ref: "#/components/parameters/QuerySearch"
        - $ref: "#/components/parameters/QueryCip"
      responses:
        "200":
          description: Réponse réussie
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/MedicamentArray"
                  - $ref: "#/components/schemas/PaginatedMedicament"
              examples:
                paginated:
                  summary: Pagination
                  value:
                    data:
                      - cis: "61504672"
                        elementPharmaceutique: "PARACETAMOL MYLAN 1 g, comprimé"
                        formePharmaceutique: "comprimé"
                        voiesAdministration: ["orale"]
                        statusAutorisation: "Autorisation active"
                        typeProcedure: "Procédure nationale"
                        etatComercialisation: "Commercialisée"
                        dateAMM: "2000-01-01"
                        titulaire: "MYLAN SAS"
                        surveillanceRenforcee: "Non"
                        composition: []
                        generiques: []
                        presentation: []
                        conditions: []
                    page: 1
                    pageSize: 10
                    totalItems: 15000
                    maxPage: 1500
                search:
                  summary: Recherche par nom
                  value:
                    - cis: "61504672"
                      elementPharmaceutique: "PARACETAMOL MYLAN 1 g, comprimé"
                      formePharmaceutique: "comprimé"
                      voiesAdministration: ["orale"]
                      statusAutorisation: "Autorisation active"
                      typeProcedure: "Procédure nationale"
                      etatComercialisation: "Commercialisée"
                      dateAMM: "2000-01-01"
                      titulaire: "MYLAN SAS"
                      surveillanceRenforcee: "Non"
                      composition: []
                      generiques: []
                      presentation: []
                      conditions: []
                    - cis: "60001234"
                      elementPharmaceutique: "DOLIPRANE 1000 mg, comprimé"
                      formePharmaceutique: "comprimé"
                      voiesAdministration: ["orale"]
                      statusAutorisation: "Autorisation active"
                      typeProcedure: "Procédure nationale"
                      etatComercialisation: "Commercialisée"
                      dateAMM: "1998-05-15"
                      titulaire: "SANOFI AVENTIS FRANCE"
                      surveillanceRenforcee: "Non"
                      composition: []
                      generiques: []
                      presentation: []
                      conditions: []
        "400":
          description: Paramètres invalides
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                bad-request:
                  value:
                    error: "Bad Request"
                    message: "Only one parameter allowed at a time"
                    code: 400
                accents-not-supported:
                  summary: Accents non supportés
                  value:
                    error: "Bad Request"
                    message: "Accents not supported. Try removing them (e.g., use 'ibuprofene' instead of 'ibuprofène')"
                    code: 400
                word-limit-exceeded:
                  summary: Limite de mots dépassée
                  value:
                    error: "Bad Request"
                    message: "search query too complex: maximum 6 words allowed"
                    code: 400
        "404":
          description: Aucun résultat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                not-found:
                  value:
                    error: "Not Found"
                    message: "No medicaments found"
                    code: 404
        "500":
          description: Erreur interne du serveur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                internal-error:
                  value:
                    error: "Internal Server Error"
                    message: "Internal server error"
                    code: 500
  /v1/medicaments/export:
    get:
      summary: Exporter tous les médicaments (v1)
      description: |
        Exporter la base de données complète de tous les médicaments.
        Utilisez ETag et Last-Modified headers pour le cache conditionnel.
      tags:
        - Médicaments (v1)
      responses:
        "200":
          description: Réponse réussie avec liste complète
          headers:
            ETag:
              schema:
                type: string
              examples:
                etag-header:
                  value: 'W/"abc123"'
            Last-Modified:
              schema:
                type: string
                format: date-time
              examples:
                last-modified-header:
                  value: "Mon, 15 Jan 2026 06:00:00 GMT"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicamentArray"
              examples:
                export:
                  summary: Export complet
                  value:
                    - cis: "61504672"
                      elementPharmaceutique: "PARACETAMOL MYLAN 1 g, comprimé"
                      formePharmaceutique: "comprimé"
                      voiesAdministration: ["orale"]
                      statusAutorisation: "Autorisation active"
                      typeProcedure: "Procédure nationale"
                      etatComercialisation: "Commercialisée"
                      dateAMM: "2000-01-01"
                      titulaire: "MYLAN SAS"
                      surveillanceRenforcee: "Non"
                      composition: []
                      generiques: []
                      presentation: []
                      conditions: []
                    - cis: "61504672"
                      elementPharmaceutique: "PARACETAMOL 500 mg, comprimé"
                      formePharmaceutique: "comprimé"
                      voiesAdministration: ["orale"]
                      statusAutorisation: "Autorisation active"
                      typeProcedure: "Procédure nationale"
                      etatComercialisation: "Commercialisée"
                      dateAMM: "2000-01-01"
                      titulaire: "MYLAN SAS"
                      surveillanceRenforcee: "Non"
                      composition: []
                      generiques: []
                      presentation: []
                      conditions: []
        "400":
          description: Code CIS invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                bad-request:
                  value:
                    error: "Bad Request"
                    message: "Invalid CIS"
                    code: 400
        "404":
          description: Médicament non trouvé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                not-found:
                  value:
                    error: "Not Found"
                    message: "Medicament not found"
                    code: 404
        "500":
          description: Erreur interne du serveur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                internal-error:
                  value:
                    error: "Internal Server Error"
                    message: "Internal server error"
                    code: 500
  /v1/medicaments/{cis}:
    get:
      summary: Obtenir un médicament par CIS (v1)
      description: |
        Récupérer un médicament spécifique par son code CIS (Code Identifiant de Spécialité).
      tags:
        - Médicaments (v1)
      parameters:
        - $ref: "#/components/parameters/PathCis"
      responses:
        "200":
          description: Médicament trouvé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicament"
              examples:
                medicament:
                  summary: Médicament trouvé
                  value:
                    cis: "61504672"
                    elementPharmaceutique: "PARACETAMOL MYLAN 1 g, comprimé"
                    formePharmaceutique: "comprimé"
                    voiesAdministration: ["orale"]
                    statusAutorisation: "Autorisation active"
                    typeProcedure: "Procédure nationale"
                    etatComercialisation: "Commercialisée"
                    dateAMM: "2000-01-01"
                    titulaire: "MYLAN SAS"
                    surveillanceRenforcee: "Non"
                    composition: []
                    generiques: []
                    presentation: []
                    conditions: []
        "400":
          description: Code CIS invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                bad-request:
                  value:
                    error: "Bad Request"
                    message: "Invalid CIS"
                    code: 400
        "404":
          description: Médicament non trouvé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                not-found:
                  value:
                    error: "Not Found"
                    message: "Medicament not found"
                    code: 404
        "500":
          description: Erreur interne du serveur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                internal-error:
                  value:
                    error: "Internal Server Error"
                    message: "Internal server error"
                    code: 500
  /v1/generiques:
    get:
      summary: Rechercher des groupes génériques par libellé (v1)
      description: |
        Recherche et récupération de groupes génériques par libellé.
        Seul un paramètre est autorisé à la fois.
      tags:
        - Génériques (v1)
      parameters:
        - $ref: "#/components/parameters/QueryLibelle"
        - $ref: "#/components/parameters/QueryGroup"
      responses:
        "200":
          description: Réponse réussie
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeneriqueListResponse"
              examples:
                search:
                  summary: Recherche par libellé
                  value:
                    - groupID: 1368
                      libelle: "PARACETAMOL 400 mg + CAFEINE 50 mg + CODEINE (PHOSPHATE DE) HEMIHYDRATE 20 mg - PRONTALGINE, comprimé"
                      medicaments:
                        - cis: "61644230"
                          elementPharmaceutique: "PRONTALGINE, comprimé"
                          formePharmaceutique: "comprimé"
                          type: "Princeps"
                          composition:
                            - elementPharmaceutique: "comprimé"
                              substance: "CAFÉINE ANHYDRE"
                              dosage: "50,0 mg"
                            - elementPharmaceutique: "comprimé"
                              substance: "PARACÉTAMOL"
                              dosage: "400,0 mg"
                            - elementPharmaceutique: "comprimé"
                              substance: "PHOSPHATE DE CODÉINE HÉMIHYDRATÉ"
                              dosage: "20,0 mg"
                        - cis: "63399979"
                          elementPharmaceutique: "PARACETAMOL/CAFEINE/CODEINE ARROW 400 mg/50 mg/20 mg, comprimé"
                          formePharmaceutique: "comprimé"
                          type: "Générique"
                          composition:
                            - elementPharmaceutique: "comprimé"
                              substance: "CAFÉINE ANHYDRE"
                              dosage: "50 mg"
                            - elementPharmaceutique: "comprimé"
                              substance: "PARACÉTAMOL"
                              dosage: "400 mg"
                            - elementPharmaceutique: "comprimé"
                              substance: "PHOSPHATE DE CODÉINE HÉMIHYDRATÉ"
                              dosage: "20 mg"
                      orphanCIS: [61586325]
                    - groupID: 1643
                      libelle: "PARACETAMOL 500 mg + CODEINE (PHOSPHATE DE) HEMIHYDRATE 30 mg - DAFALGAN CODEINE, comprimé pelliculé"
                      medicaments:
                        - cis: "66003374"
                          elementPharmaceutique: "DAFALGAN CODEINE, comprimé pelliculé"
                          formePharmaceutique: "comprimé pelliculé"
                          type: "Princeps"
                          composition:
                            - elementPharmaceutique: "comprimé pelliculé"
                              substance: "PARACÉTAMOL"
                              dosage: "500 mg"
                            - elementPharmaceutique: "comprimé pelliculé"
                              substance: "CODÉINE (PHOSPHATE DE) HÉMIHYDRATÉ"
                              dosage: "30 mg"
                        - cis: "69458587"
                          elementPharmaceutique: "PARACETAMOL/CODEINE BIOGARAN 500 mg/30 mg, comprimé"
                          formePharmaceutique: "comprimé"
                          type: "Générique"
                          composition:
                            - elementPharmaceutique: "comprimé"
                              substance: "PARACÉTAMOL"
                              dosage: "500,00 mg"
                            - elementPharmaceutique: "comprimé"
                              substance: "PHOSPHATE DE CODÉINE HÉMIHYDRATÉ"
                              dosage: "30,00 mg"
                      orphanCIS: null
        "400":
          description: Paramètres invalides
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                bad-request:
                  value:
                    error: "Bad Request"
                    message: "Needs libelle param"
                    code: 400
        "404":
          description: Aucun générique trouvé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                not-found:
                  value:
                    error: "Not Found"
                    message: "No generiques found"
                    code: 404
        "500":
          description: Erreur interne du serveur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                internal-error:
                  value:
                    error: "Internal Server Error"
                    message: "Internal server error"
                    code: 500
  /v1/presentations/{cip}:
    get:
      summary: Obtenir des présentations (v1)
      description: |
        Recherche de présentations par code CIP (CIP7 ou CIP13).
      tags:
        - Présentations (v1)
      parameters:
        - $ref: "#/components/parameters/PathCip"
      responses:
        "200":
          description: Réponse réussie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresentationResponse"
              examples:
                presentation:
                  value:
                    cis: "60904643"
                    cip7: "2756239"
                    libelle: "CODOLIPRANE 500 mg/30 mg, comprimé (16)"
                    statusAdministratif: "Présentation active"
                    etatComercialisation: "Commercialisée"
                    dateDeclaration: "1965-01-19"
                    cip13: "3400936403114"
                    agreement: "non"
                    tauxRemboursement: "65%"
                    prix: 3.85
        "400":
          description: Code CIP invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                bad-request:
                  value:
                    error: "Bad Request"
                    message: "Invalid CIP"
                    code: 400
        "404":
          description: Présentation non trouvée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                not-found:
                  value:
                    error: "Not Found"
                    message: "Presentation not found"
                    code: 404
        "500":
          description: Erreur interne du serveur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                internal-error:
                  value:
                    error: "Internal Server Error"
                    message: "Internal server error"
                    code: 500
  /v1/generiques/{groupID}:
    get:
      summary: Obtenir un groupe générique par ID (v1)
      description: |
        Récupérer un groupe générique spécifique par son ID.
        Utilisez ETag et Last-Modified headers pour le cache conditionnel.
      tags:
        - Génériques (v1)
      parameters:
        - $ref: "#/components/parameters/PathGroupID"
      responses:
        "200":
          description: Groupe générique trouvé
          headers:
            ETag:
              schema:
                type: string
              examples:
                etag-header:
                  value: 'W/"abc123"'
            Last-Modified:
              schema:
                type: string
                format: date-time
              examples:
                last-modified-header:
                  value: "Mon, 15 Jan 2026 06:00:00 GMT"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneriqueListResponse"
              examples:
                generique:
                  summary: Groupe générique trouvé
                  value:
                    groupID: 1643
                    libelle: "PARACETAMOL 500 mg + CODEINE (PHOSPHATE DE) HEMIHYDRATE 30 mg - DAFALGAN CODEINE, comprimé pelliculé"
                    medicaments:
                      - cis: "66003374"
                        elementPharmaceutique: "DAFALGAN CODEINE, comprimé pelliculé"
                        formePharmaceutique: "comprimé pelliculé"
                        type: "Princeps"
                        composition: []
                    orphanCIS: null
        "400":
          description: ID de groupe invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                bad-request:
                  value:
                    error: "Bad Request"
                    message: "Invalid group ID"
                    code: 400
        "404":
          description: Groupe générique non trouvé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                not-found:
                  value:
                    error: "Not Found"
                    message: "Generique group not found"
                    code: 404
        "500":
          description: Erreur interne du serveur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                internal-error:
                  value:
                    error: "Internal Server Error"
                    message: "Internal server error"
                    code: 500
  /v1/diagnostics:
    get:
      summary: Obtenir les diagnostics système détaillés (v1)
      description: |
        Endpoint de monitoring avancé fournissant des métriques système détaillées et un rapport d'intégrité des données.

        **Informations fournies :**
        - Métriques système : uptime, mémoire, goroutines
        - Intégrité des données : médicaments sans conditions, génériques, présentations, compositions
        - Métriques de données : âge des données, prochaine mise à jour planifiée

        Cache de 10 secondes pour éviter le spam tout en gardant les données fraîches.
      tags:
        - Système
      responses:
        "200":
          description: Diagnostics système
          headers:
            Cache-Control:
              schema:
                type: string
              examples:
                cache-header:
                  value: "public, max-age=10"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticsResponse"
              examples:
                diagnostics:
                  summary: Diagnostics système complets
                  value:
                    timestamp: "2026-01-15T14:30:45Z"
                    uptime_seconds: 86400
                    next_update: "2026-01-15T18:00:00Z"
                    data_age_hours: 2.5
                    system:
                      goroutines: 45
                      memory:
                        alloc_mb: 130
                        sys_mb: 150
                        num_gc: 25
                    data_integrity:
                      medicaments_without_conditions:
                        count: 1250
                        sample_cis: [60012345, 60023456]
                      medicaments_without_generiques:
                        count: 8450
                        sample_cis: [61504672, 61001234]
                      medicaments_without_presentations:
                        count: 3200
                        sample_cis: [62003456, 62004567]
                      medicaments_without_compositions:
                        count: 890
                        sample_cis: [63005678, 63006789]
                      generique_only_cis:
                        count: 45
                        sample_cis: [64007890, 64008901]
        "500":
          description: Erreur interne du serveur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                internal-error:
                  value:
                    error: "Internal Server Error"
                    message: "Internal server error"
                    code: 500
...
components:
  parameters:
    QueryPage:
      name: page
      in: query
      required: false
      description: Numéro de page (pagination à 10 éléments par page)
      schema:
        type: integer
        minimum: 1
    QuerySearch:
      name: search
      in: query
      required: false
      description: |
        Terme de recherche multi-mots (1-6 mots requis, logique ET).
        Les mots sont séparés par + ou espace dans l'URL.
        Minimum 3 caractères, maximum 50 caractères.
      schema:
        type: string
        minLength: 3
        maxLength: 50
        pattern: "^[a-zA-Z0-9\\s\\-\\.\\+']+$"
    QueryCip:
      name: cip
      in: query
      required: false
      description: Code CIP (CIP7 ou CIP13)
      schema:
        type: string
        pattern: "^[0-9]{7}|[0-9]{13}$"
...
    QueryGroup:
      name: group
      in: query
      required: false
      description: ID de groupe (entier positif)
      schema:
        type: integer
        minimum: 1
        maximum: 99999
    PathCis:
      name: cis
      in: path
      required: true
      description: Code CIS
      schema:
        type: string
        pattern: "^[0-9]{1,9}$"
...
  schemas:
    PaginatedMedicament:
      type: object
      title: PaginatedMedicament
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Medicament"
          title: Données des médicaments
        page:
          type: integer
          title: Numéro de page
        pageSize:
          type: integer
          title: Taille de page
        totalItems:
          type: integer
          title: Nombre total d'éléments
        maxPage:
          type: integer
          title: Page maximum
...
    ErrorResponse:
      type: object
      title: ErrorResponse
      properties:
        error:
          type: string
          title: Statut d'erreur
        message:
          type: string
          title: Message d'erreur
        code:
          type: integer
          title: Code de statut HTTP
    MedicamentArray:
      type: array
      title: MedicamentArray
      items:
        $ref: "#/components/schemas/Medicament"
...